Hide implementation-specific wg stats parsers
